# Speed up `connected_components` and `weakly_connected_components`