## Potential Speed up `connected_components` and `weakly_connected_components`

This PR introduces lightweight optimizations to several connectivity-checking functions by adding early returns based on simple graph invariants. These heuristics help short-circuit computation in cases where connectivity can be ruled out without traversing the graph.

### Summary of changes:
- Adjusted BFS calls in `connected_components` and `weakly_connected_components` to reduce traversal overhead when many nodes have already been seen.

These changes preserve correctness while reducing unnecessary work, especially on dense graphs.

### Performance improvements results

The optimization shows the most significant gains on disconnected graphs with large components, achieving up to a 367× speed-up for n=1000. This case is artificially generated to make current version need to visit the same nodes in a connected component multiple times. These are the types of cases where optimization works best.

For connected and random graphs, performance changes are generally smaller, often within a ±10% range. In a few low-n random graph cases, the optimized version was marginally slower, likely due to overhead not offset by structural gains. 

Overall, the optimization is highly effective in worst-case disconnected scenarios while maintaining neutral or slightly improved performance elsewhere.

#### Single Node + Complete Component (Disconnected, 100 runs)

| n    | Without Optimization (s) | With Optimization (s) | Speed-up |
|------|---------------------------|------------------------|----------|
| 10   | 0.000008                  | 0.000004               | 2.00× (faster) |
| 100  | 0.000552                  | 0.000017               | 32.47× (faster) |
| 1000 | 0.049941                  | 0.000136               | 367.21× (faster) |


#### Random Graphs (p=0.01, 10000 runs each)

| n    | Seed | Without Optimization (s) | With Optimization (s) | Speed-up |
|------|------|---------------------------|------------------------|----------|
| 10   | 42   | 0.000008                  | 0.000015               | 0.53× (slower) |
| 10   | 99   | 0.000007                  | 0.000008               | 0.88× (slower) |
| 10   | 123  | 0.000007                  | 0.000009               | 0.78× (slower) |
| 100  | 42   | 0.000048                  | 0.000047               | 1.02× (faster) |
| 100  | 99   | 0.000047                  | 0.000053               | 0.89× (slower) |
| 100  | 123  | 0.000051                  | 0.000049               | 1.04× (faster) |
| 1000 | 42   | 0.000597                  | 0.000577               | 1.03× (faster) |
| 1000 | 99   | 0.000785                  | 0.000733               | 1.07× (faster) |
| 1000 | 123  | 0.000688                  | 0.000614               | 1.12× (faster) |

#### Connected Graphs (Complete Graphs, 10000 runs)

| n    | Without Optimization (s) | With Optimization (s) | Speed-up |
|------|---------------------------|------------------------|----------|
| 10   | 0.000003                  | 0.000003               | 1.00× (same) |
| 100  | 0.000018                  | 0.000016               | 1.13× (faster) |
| 1000 | 0.000149                  | 0.000142               | 1.05× (faster) |

#### Benchmark code

```python

import networkx as nx
import timeit

def measure_runtime(func, runs=100):
    """Measure average runtime over multiple runs."""
    timer = timeit.Timer(func)
    times = timer.repeat(repeat=1, number=runs)
    return times[0] / runs

def random_graph_test(ns, p, seeds, runs=100):
    print(f"\n=== Random Graphs (p={p}, {runs} runs each) ===")
    for n in ns:
        for seed in seeds:
            G = nx.erdos_renyi_graph(n, p, seed=seed)
            def run():
                list(nx.connected_components(G))
            t = measure_runtime(run, runs)
            print(f"n={n}, seed={seed}: {t:.6f} seconds")

def connected_graph_test(ns, runs=100):
    print(f"\n=== Connected Graphs (Complete Graphs, {runs} runs) ===")
    for n in ns:
        G = nx.complete_graph(n)
        def run():
            list(nx.connected_components(G))
        t = measure_runtime(run, runs)
        print(f"n={n}: {t:.6f} seconds")

def single_and_complete_component_test(ns, runs=100):
    print(f"\n=== Single Node + Complete Component (Disconnected, {runs} runs) ===")
    for n in ns:
        G = nx.Graph()
        G.add_node(0)
        G.add_edges_from((i, j) for i in range(1, n) for j in range(i + 1, n))
        def run():
            list(nx.connected_components(G))
        t = measure_runtime(run, runs)
        print(f"n={n}: {t:.6f} seconds")

if __name__ == "__main__":
    Ns = [10, 100, 1000]
    P = 0.01
    SEEDS = [42, 99, 123]

    random_graph_test(Ns, P, SEEDS, runs=10_000)
    connected_graph_test(Ns, runs=10_000)
    single_and_complete_component_test(Ns, runs=100)
```