@dschult 2025-04-14
There is *some* computational cost to networkx counting the number of edges. It is about a third the time of using `bfs_edges` for 1000 nodes. And it should scale with number of nodes instead of number of edges.So I think this should be effective as a shortcut approach. 

Are there cases when a user would probably not want these shortcuts computed? For the case of 1000 nodes, I estimate it would take 465us instead of 350us (~30% longer) for a non-shortcut result. And 115us for the shortcut case. But those estimates are not using this code.

If that's an issue we could add a kwarg to turn on (or off) the shortcut feature. Thoughts?