@amcandio 2025-04-14
Ah, I did not know that! I suspected len was expensive after seeing some comments in the code. What is the historical context for those fields to not be cached?