@dschult 2025-04-14
`len` is typically fast -- so `len(G)` is too. But we have to compute the `len` of each nbrdict in the dict-of-dict data structure. I think of it as being quick O(n) -- but then I think of `is_connected` as being quick too O(n+m) = O(m) :)

In terms of caching, we thought about caching way back in the day -- and some factors led us not to choose it: 
- caching updates do slow down building and changing the graph and many algorithms/codepaths never use it
- the data structure could be changed directly by users without our knowing about it (not so true anymore -- views kind of discourage that kind of grow-your-own approach). 
- there is a simplicity to the data structure without caching (again views detract from that argument
- it seemed at the time, that the user could easily cache the result themselves and use it where needed. 

Another data structure could have avoided this issue -- if we had the dict-of-dict point to the relevant edge info within a list of edge info dicts. Then the number of (unweighted) edges would be fast: `len(edge_info_list)`.

That's the history of it -- but the history leading to no caching doesn't mean we can't start caching now -- of course, we should take into account these factors in our decision.

I did a little more testing, and it looks like for `G=nx.binomial_graph(10_000, 0.0001)` that `%timeit G.number_of_edges()` gives 1.12ms while `%timeit nx.is_connected(G)` takes 21us.  So there must be features of the code that make these large graphs with few edges fast already. Have you done some tests with this PR? Which cases lead to improvements?