@amcandio 2025-04-15
Thanks for the context! I believe adding caching would unblock many optimizations like this.

> So there must be features of the code that make these large graphs with few edges fast already

Yes, I can see why now. It's because DFS is exiting early while edge counting needs to visit all edges.

> Have you done some tests with this PR? Which cases lead to improvements?