@amcandio 2025-04-30
Hey @dschult I got time to come back to this. I removed the number of edges optimizations because they were actually slower, leaving the PR with only the BFS stop-early optimization.

I'm seeing very good results on graphs where one big component is explored last. For the rest of tested graphs, I see small mixed results. What do you think?